repeatable builds
we now have no fluctuation in the order of menu items anymore.
running the build twice and comparing it via `diff -r` yields only a
single CSS-file (which we'll sort out later). this way we can change
tooling while being sure nothing broke.
